implement completion strategy without task list client, fix #41
