+++ engine-adapter/camunda-platform-8-core/src/main/kotlin/dev/bpmcrafters/processengineapi/adapter/c8/task/completion/C8ZeebeUserTaskCompletionApiImpl.kt
@@ -1,0 +1,42 @@
+package dev.bpmcrafters.processengineapi.adapter.c8.task.completion
+
+import dev.bpmcrafters.processengineapi.Empty
+import dev.bpmcrafters.processengineapi.adapter.commons.task.SubscriptionRepository
+import dev.bpmcrafters.processengineapi.task.CompleteTaskByErrorCmd
+import dev.bpmcrafters.processengineapi.task.CompleteTaskCmd
+import dev.bpmcrafters.processengineapi.task.UserTaskCompletionApi
+import io.camunda.zeebe.client.ZeebeClient
+import mu.KLogging
+import java.util.concurrent.CompletableFuture
+import java.util.concurrent.Future
+
+class C8ZeebeUserTaskCompletionApiImpl(
+  private val zeebeClient: ZeebeClient,
+  private val subscriptionRepository: SubscriptionRepository
+) : UserTaskCompletionApi {
+
+  companion object : KLogging()
+
+  override fun completeTask(cmd: CompleteTaskCmd): Future<Empty> {
+    zeebeClient
+      .newCompleteCommand(cmd.taskId.toLong())
+      .variables(cmd.get())
+      .send()
+    subscriptionRepository.deactivateSubscriptionForTask(cmd.taskId)?.apply {
+      termination.accept(cmd.taskId)
+    }
+    return CompletableFuture.completedFuture(Empty)
+  }
+
+  override fun completeTaskByError(cmd: CompleteTaskByErrorCmd): Future<Empty> {
+    zeebeClient
+      .newThrowErrorCommand(cmd.taskId.toLong())
+      .errorCode(cmd.errorCode)
+      .variables(cmd.get())
+      .send()
+    subscriptionRepository.deactivateSubscriptionForTask(cmd.taskId)?.apply {
+      termination.accept(cmd.taskId)
+    }
+    return CompletableFuture.completedFuture(Empty)
+  }
+}
