+++ engine-adapter/camunda-platform-8-spring-boot-starter/src/main/kotlin/dev/bpmcrafters/processengineapi/adapter/c8/springboot/C8AdapterAutoConfiguration.kt
@@ -70,55 +70,4 @@ class C8AdapterAutoConfiguration {
 
   @Bean
   fun subscriptionRepository(): SubscriptionRepository = InMemSubscriptionRepository()
-
-  @Bean
-  fun externalTaskCompletionStrategy(
-    zeebeClient: ZeebeClient,
-    subscriptionRepository: SubscriptionRepository,
-  ): ExternalTaskCompletionApi =
-    C8ZeebeExternalServiceTaskCompletionApiImpl(
-      zeebeClient = zeebeClient,
-      subscriptionRepository = subscriptionRepository
-    )
-
-
-  @Bean
-  @ConditionalOnMissingBean
-  fun camundaTaskListClientSaaS(
-    zeebeClientCloudConfigurationProperties: ZeebeClientConfigurationProperties,
-    c8AdapterProperties: C8AdapterProperties
-  ): CamundaTaskListClient {
-    /*
-    val jwtConfig = JwtConfig()
-    jwtConfig.addProduct(Product.TASKLIST,
-      JwtCredential(
-        zeebeClientCloudConfigurationProperties.cloud.clientId,
-        zeebeClientCloudConfigurationProperties.cloud.clientSecret,
-        "tasklist.camunda.io",
-        "https://login.cloud.camunda.io/oauth/token")
-    )
-    val authentication = SaaSAuthentication.builder().jwtConfig(jwtConfig).build()
-    */
-
-    return CamundaTaskListClient
-      .builder()
-      .taskListUrl(c8AdapterProperties.userTasks.tasklistUrl)
-      .saaSAuthentication(
-        zeebeClientCloudConfigurationProperties.cloud.clientId,
-        zeebeClientCloudConfigurationProperties.cloud.clientSecret,
-      )
-      .shouldReturnVariables()
-      // .authentication(authentication) // produces NPE
-      .build()
-  }
-
-  @Bean
-  fun userTaskCompletionStrategy(
-    taskListClient: CamundaTaskListClient,
-    subscriptionRepository: SubscriptionRepository
-  ): UserTaskCompletionApi =
-    C8TaskListClientUserTaskCompletionApiImpl(
-      taskListClient = taskListClient,
-      subscriptionRepository = subscriptionRepository
-    )
 }
