+++ engine-adapter/camunda-platform-8-spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,3 @@
 dev.bpmcrafters.processengineapi.adapter.c8.springboot.C8AdapterAutoConfiguration
 dev.bpmcrafters.processengineapi.adapter.c8.springboot.DeliveryAutoConfiguration
+dev.bpmcrafters.processengineapi.adapter.c8.springboot.CompletionAutoConfiguration
