+++ engine-adapter/camunda-platform-8-spring-boot-starter/src/test/resources/application-itest.yml
@@ -12,6 +12,7 @@ dev:
             fixed-rate-schedule-rate: 100000
             tasklist-url: localhost
             fixed-rate-refresh-rate: 100000
+            completion-strategy: tasklist
           service-tasks:
             delivery-strategy: subscription
             worker-id: execute-action-external
