+++ examples/java-c8/src/main/resources/application-saas.yml
@@ -9,6 +9,7 @@ dev:
         c8:
           user-tasks:
             delivery-strategy: subscription_refreshing
+            completion-strategy: job
             fixed-rate-schedule-rate: 2000 # every 2 seconds
             tasklist-url: https://${zeebe.client.cloud.region}.tasklist.camunda.io/${zeebe.client.cloud.clusterId}
             fixed-rate-refresh-rate: 2000 # every 2 seconds
